Cambios y fixes menores en los daos, en el SalesAccountant y en el BakeryPhoneOp para que funcione la interacción con la BD

+++ src/model/SalesAccountant.java
@@ -41,7 +41,7 @@ public class SalesAccountant {
             
             //adjust the price based on the product quantity
             double adjustedPrice = unitaryPrice * 
-                (double) mod_ordersList.getValueAt( productCount, OrdersList.PRODUCT_QUANTITY );
+                (int) mod_ordersList.getValueAt( productCount, OrdersList.PRODUCT_QUANTITY );
             
             //write the adjusted price in the orders list
             mod_ordersList.setValueAt( adjustedPrice, productCount, OrdersList.PRODUCT_PRICE );
